Avoid compile warnings.

+++ src/kde/sambe.c
@@ -15,7 +15,7 @@ size_t g_numXs;
 
 gsl_vector* g_movedPattern;
 
-int g_k;
+size_t g_k;
 gsl_matrix* g_distanceMatrix;
 gsl_matrix* g_nearestNeighbours;
 
@@ -68,7 +68,7 @@ double finalDensitySinglePattern(gsl_vector *x, size_t xIdx) {
 
 void determineGlobalKernelShape(size_t patternIdx) {
     /* Compute K nearest neighbours */
-    computeKNearestNeighbours(g_k, (int) patternIdx,
+    computeKNearestNeighbours(g_k, patternIdx,
                               g_xs, g_distanceMatrix,
                               g_nearestNeighbours);
 
@@ -89,7 +89,7 @@ gsl_vector *subtract(gsl_vector *termA, gsl_vector *termB, gsl_vector *result) {
     return result;
 }
 
-void allocateGlobals(size_t dataDimension, size_t num_xi_s, int k) {
+void allocateGlobals(size_t dataDimension, size_t num_xi_s, size_t k) {
     g_globalBandwidthMatrix = gsl_matrix_alloc(dataDimension, dataDimension);
     g_distanceMatrix = gsl_matrix_alloc(num_xi_s, num_xi_s);
     g_nearestNeighbours = gsl_matrix_alloc(k, dataDimension);
@@ -111,7 +111,7 @@ void prepareGlobals(gsl_matrix *xs,
     g_localBandwidths = localBandwidths;
     g_globalBandwidth = globalBandwidth;
     g_kernel = kernel;
-    g_k = k;
+    g_k = (size_t) k;
 
     g_numXs = xs->size1;
